Updating packages to follow python convention
