+++ sslt/transforms/transform.py
@@ -1,0 +1,42 @@
+
+from typing import Any, List
+
+
+class _Transform:
+    """This class is a base class for all transforms. Transforms is just a 
+    fancy word for a function that takes an input and returns an output. The 
+    input and output can be anything. However, transforms operates over a 
+    single sample of data and does not require any additional information to
+    perform the transformation. The __call__ method should be overridden in
+    subclasses to define the transformation logic.
+    """
+
+    def __call__(self, *args, **kwargs) -> Any:
+        """Implement the transformation logic in this method. Usually, the 
+        transformation is applyied on a single sample of data.
+        """
+        raise NotImplementedError
+
+
+class TransformPipeline(_Transform):
+    """Apply a sequence of transforms to a single sample of data and return the 
+    transformed data.
+    """
+    def __init__(self, transforms: List[_Transform]):
+        """Apply a sequence of transforms to a single sample of data and return
+        the transformed data.
+
+        Parameters
+        ----------
+        transforms : List[_Transform]
+            A list of transforms to be applied to the input data.
+        """
+        self.transforms = transforms
+
+    def __call__(self, x: Any) -> Any:
+        """Apply a sequence of transforms to a single sample of data and return
+        the transformed data.
+        """
+        for transform in self.transforms:
+            x = transform(x)
+        return x
